+++ GnugatRedaktiloBundle.php
@@ -11,9 +11,19 @@
 
 namespace Gnugat\RedaktiloBundle;
 
+use Gnugat\RedaktiloBundle\DependencyInjection\CommandCompilerPass;
+use Gnugat\RedaktiloBundle\DependencyInjection\SearchStrategyCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class GnugatRedaktiloBundle extends Bundle
 {
+    /**
+     * {@inheritDoc}
+     */
+    public function build(ContainerBuilder $container)
+    {
+        $container->addCompilerPass(new SearchStrategyCompilerPass());
+        $container->addCompilerPass(new CommandCompilerPass());
+    }
 }