+++ Tests/app/autoload.php
@@ -1,0 +1,13 @@
+<?php
+
+/*
+ * This file is part of the RedaktiloBundle project.
+ *
+ * (c) Loïc Chardonnet <loic.chardonnet@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+$loader = require __DIR__.'/../../vendor/autoload.php';
+require __DIR__.'/AppKernel.php';