+++ CHANGELOG.md
@@ -2,4 +2,12 @@
 
 ## 1.0.0
 
+* added `redaktilo.search_strategy` tag
+* added `redaktilo.command` tag
+* added `redaktilo.text_factory` service
+* added `redaktilo.editor` service
+* added `gnugat/redaktilo` v1.2
+* added `symfony/yaml` v2.3
+* added `symfony/config` v2.3
+* added `symfony/dependency-injection` v2.3
 * added `symfony/http-kernel` v2.3